+++ client/v2/autocli/app.go
@@ -9,14 +9,12 @@ import (
 	autocliv1 "cosmossdk.io/api/cosmos/autocli/v1"
 	"cosmossdk.io/client/v2/autocli/flag"
 	"cosmossdk.io/client/v2/autocli/keyring"
-	"cosmossdk.io/core/address"
 	"cosmossdk.io/core/appmodule"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/runtime"
 )
 
 // AppOptions are autocli options for an app. These options can be built via depinject based on an app config. Ex:
@@ -42,10 +40,8 @@ type AppOptions struct {
 	// module or need to be improved.
 	ModuleOptions map[string]*autocliv1.ModuleOptions `optional:"true"`
 
-	// AddressCodec is the address codec to use for the app.
-	AddressCodec          address.Codec
-	ValidatorAddressCodec runtime.ValidatorAddressCodec
-	ConsensusAddressCodec runtime.ConsensusAddressCodec
+	// ClientCtx contains the necessary information needed to execute the commands.
+	ClientCtx *client.Context
 
 	// Keyring is the keyring to use for client/v2.
 	Keyring keyring.Keyring `optional:"true"`
@@ -70,12 +66,10 @@ func (appOptions AppOptions) EnhanceRootCommand(rootCmd *cobra.Command) error {
 	builder := &Builder{
 		Logger: appOptions.Logger,
 		Builder: flag.Builder{
-			TypeResolver:          protoregistry.GlobalTypes,
-			FileResolver:          proto.HybridResolver,
-			AddressCodec:          appOptions.AddressCodec,
-			ValidatorAddressCodec: appOptions.ValidatorAddressCodec,
-			ConsensusAddressCodec: appOptions.ConsensusAddressCodec,
-			Keyring:               appOptions.Keyring,
+			TypeResolver: protoregistry.GlobalTypes,
+			FileResolver: proto.HybridResolver,
+			ClientCtx:    appOptions.ClientCtx,
+			Keyring:      appOptions.Keyring,
 		},
 		GetClientConn: func(cmd *cobra.Command) (grpc.ClientConnInterface, error) {
 			return client.GetClientQueryContext(cmd)