+++ x/staking/types/keys_test.go
@@ -1,11 +1,9 @@
 package types_test
 
 import (
-	"bytes"
 	"encoding/hex"
 	"math/big"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/require"
 
@@ -46,30 +44,4 @@ func TestGetValidatorPowerRank(t *testing.T) {
 
 		require.Equal(t, tt.wantHex, got, "Keys did not match on test case %d", i)
 	}
-}
-
-func TestGetValidatorQueueKey(t *testing.T) {
-	ts := time.Now()
-	height := int64(1024)
-
-	bz := types.GetValidatorQueueKey(ts, height)
-	rTs, rHeight, err := types.ParseValidatorQueueKey(bz)
-	require.NoError(t, err)
-	require.Equal(t, ts.UTC(), rTs.UTC())
-	require.Equal(t, rHeight, height)
-}
-
-func TestTestGetValidatorQueueKeyOrder(t *testing.T) {
-	ts := time.Now().UTC()
-	height := int64(1000)
-
-	endKey := types.GetValidatorQueueKey(ts, height)
-
-	keyA := types.GetValidatorQueueKey(ts.Add(-10*time.Minute), height-10)
-	keyB := types.GetValidatorQueueKey(ts.Add(-5*time.Minute), height+50)
-	keyC := types.GetValidatorQueueKey(ts.Add(10*time.Minute), height+100)
-
-	require.Equal(t, -1, bytes.Compare(keyA, endKey)) // keyA <= endKey
-	require.Equal(t, -1, bytes.Compare(keyB, endKey)) // keyB <= endKey
-	require.Equal(t, 1, bytes.Compare(keyC, endKey))  // keyB >= endKey
 }